requirements.txt
Flask
cohere
Flask-CORS
python-dotenv
pinecone-client
langchain-cohere
langchain-pinecone
langchain-core
Gunicorn